[#4] Separate login form into a child widget to reduce the complexity
